Moved navbar related components from page folder to components folder
