+++ src/pages/home/HomePageLayout.jsx
@@ -1,11 +1,10 @@
 import {Box, Grid2, useTheme} from "@mui/material";
 import Grid from "@mui/material/Grid2";
-import ResponsiveAppBar from "../../components/ResponsiveAppBar.jsx";
+import ResponsiveAppBar from "../../components/ResponsiveAppBar/ResponsiveAppBar.jsx";
 
 const HomePageLayout = () => {
     //Import the theme that was provided in the HomePage.jsx component
     const theme = useTheme();
-
 
     return (
         <Box style={{height: "100vh", width: "100vw", background: `linear-gradient( ${theme.palette.primary.main},${theme.palette.secondary.main} )`}} sx={{padding: "16px"}} elevation={4}>
